pull classes into their own mini-libs
